refactored tasks in ContentScannerController
