+++ src/main/java/blackmediamanager/application/controller/ContentScannerAppController.java
@@ -1,0 +1,5 @@
+package blackmediamanager.application.controller;
+
+public class ContentScannerAppController {
+
+}
